pyproject.toml
[project]
name = "dormer"
version = "0.1"
authors = [
    {name = "Tom Parker-Shemilt", email = "palfrey@tevp.net"},
]
description = "Tool for saving/restoring i3 workspace->output mappings"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "AGPL3"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "i3ipc",
    "pyyaml",
]

[project.urls]
Homepage = "https://github.com/palfrey/dormer"

[tool.setuptools.packages.find]
include = ["dormer"]

[project.scripts]
dormer = "dormer:main"

[[tool.mypy.overrides]]
module = 'i3ipc'
ignore_missing_imports = true # FIXME: Blocked on https://github.com/altdesktop/i3ipc-python/pull/205

[[tool.mypy.overrides]]
module = 'pyfakefs.fake_filesystem'
ignore_missing_imports = true # FIXME: See https://github.com/pytest-dev/pyfakefs/issues/599

[tool.coverage.run]
omit = ["test_*"]